src/client/python/requirements.txt
pyyaml
pyzmq